Version 0.0.1.0
Created an custom connection manager (DBConnMan) where it manages the connection and the querys to the database. See file DBConnManager.php.

Added parsers for Company table and Shareholders table.

Created basic messages to be printed by any system for debuging.

Added basic connection and database data to an seperated file.

Updated the installation files with the new object DBConnMan class.