Version 0.0.1.8
Added basic parsers to the database for the communication layer.

Changed positions of some files to organize the project.

Corrected some errors in the database where some tables had missing rows and corrected some errors in table view Company.

Added the apropriated data to the DemoData.php to fill the new columns.

Updated the DBConnManager class to include support for retrieving data from the database.