version 0.0.5.0
Changed the logic structure of the pages and added new forms.

Now every new html are called from the content component and every overview is considered as modules.

Every time a developer needs to show data or any other kind of html in the page, the content area is created for this purpose and it is adaptable to every html needs.

for development purposes the access level for the sessions have been changed from 1 access to 0 access.

The database follows a waterfall access level, zero is the highest level of security access.

Now queries only retrieve the data of the appropriate access level of the session.